neuro_run_project/requirements.txt
google-generativeai
chromadb
python-dotenv
